Converted to support .Net Framework and .Net Standard
Converted to support both .Net Framework 4.7.1 and .Net Standard 2.0.
CI: Replaced test connection string with environmental variable, added OneTimeSetup and OneTimeTeardown for the test scenario.
