--- Frends.Community.Postgre.Tests/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit" version="3.11.0" targetFramework="net452" />
-  <package id="NUnit3TestAdapter" version="3.12.0" targetFramework="net452" />
-</packages>
