--- Frends.Community.Postgre.nuspec
@@ -1,27 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2011/10/nuspec.xsd">
-  <metadata>
-    <id>Frends.Community.Postgre</id>
-    <version>1.2.0</version>
-    <title>Frends.Community.Postgre</title>
-    <authors>HiQ Finland</authors>
-    <owners />
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Frends.Community.Postgre</description>
-    <summary />
-    <dependencies>
-      <dependency id="System.Threading.Tasks.Extensions" version="4.5.2" />
-      <dependency id="Npgsql" version="4.0.5" />
-      <dependency id="Newtonsoft.Json" version="12.0.1"/>
-    </dependencies>
-      <frameworkAssemblies>
-      <frameworkAssembly assemblyName="System" />
-      <frameworkAssembly assemblyName="System.Data" />
-    </frameworkAssemblies>
-  </metadata>
-  <files>
-    <file src="Frends.Community.Postgre\bin\Release\Frends.Community.Postgre.dll" target="lib\net452\Frends.Community.Postgre.dll" />
-    <file src="Frends.Community.Postgre\bin\Release\Frends.Community.Postgre.xml" target="lib\net452\Frends.Community.Postgre.xml" />
-    <file src="FrendsTaskMetadata.json" target="FrendsTaskMetadata.json"/>
-  </files>
-</package>
