+++ Frends.Community.Postgre.sln
@@ -1,17 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27004.2010
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29613.14
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FRENDS.Community.Postgre", "Frends.Community.Postgre\FRENDS.Community.Postgre.csproj", "{30071232-AFDB-4AC4-A48D-BCF9D55CDAD1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Frends.Community.Postgre", "Frends.Community.Postgre\Frends.Community.Postgre.csproj", "{35C305C0-8108-4A98-BB1D-AFE5C926239E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.Postgre.Tests", "Frends.Community.Postgre.Tests\Frends.Community.Postgre.Tests.csproj", "{2AFB9C15-D942-4288-BC36-C4342D458650}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Frends.Community.Postgre.Tests", "Frends.Community.Postgre.Tests\Frends.Community.Postgre.Tests.csproj", "{8CA92187-8E4F-4414-803B-EC899479022E}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{AD988466-462E-48C9-9268-8F420E340A67}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{78F7F22E-6E20-4BCE-8362-0C558568B729}"
 	ProjectSection(SolutionItems) = preProject
-		Frends.Community.Postgre.nuspec = Frends.Community.Postgre.nuspec
-		FrendsTaskMetadata.json = FrendsTaskMetadata.json
-		LICENSE.txt = LICENSE.txt
+		.gitignore = .gitignore
 		README.md = README.md
 	EndProjectSection
 EndProject
@@ -21,19 +19,19 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{30071232-AFDB-4AC4-A48D-BCF9D55CDAD1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{30071232-AFDB-4AC4-A48D-BCF9D55CDAD1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{30071232-AFDB-4AC4-A48D-BCF9D55CDAD1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{30071232-AFDB-4AC4-A48D-BCF9D55CDAD1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2AFB9C15-D942-4288-BC36-C4342D458650}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2AFB9C15-D942-4288-BC36-C4342D458650}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2AFB9C15-D942-4288-BC36-C4342D458650}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2AFB9C15-D942-4288-BC36-C4342D458650}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A838E395-A572-40D3-81E8-7332728C6EF2}
+		SolutionGuid = {55BC6629-85C9-48D8-8CA2-B0046AF1AF4B}
 	EndGlobalSection
 EndGlobal
