+++ Frends.Community.Postgre/Definition.cs
@@ -1,4 +1,7 @@
-﻿using System.ComponentModel;
+﻿#pragma warning disable 1591
+
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 
 namespace Frends.Community.Postgre
 {
