+++ Frends.Community.Postgre/Extensions.cs
@@ -1,10 +1,11 @@
-﻿using System;
-using Newtonsoft.Json.Linq;
-using System.Xml;
-using System.Text;
+﻿using Newtonsoft.Json.Linq;
 using Npgsql;
+using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
+using System.Text;
+using System.Xml;
 
 namespace Frends.Community.Postgre
 {
@@ -75,4 +76,6 @@ namespace Frends.Community.Postgre
             }
         }
     }
+
 }
+
