--- Frends.Community.Postgre/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net452" />
-  <package id="Npgsql" version="4.0.4" targetFramework="net452" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net452" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.2" targetFramework="net452" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net452" />
-</packages>
