+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2017 Community HiQ
+Copyright (c) 2020 Community HiQ
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
