+++ README.md
@@ -1,80 +1,75 @@
-- [Frends.Community.Files.Postgre](#frendscommunitypostgre)
-   - [Installing](#installing)
-   - [Tasks](#tasks)
-      - [QueryData](#querydata)
-         - [Input](#input)
-         - [ConnectionInformation](#connectioninformation)
-         - [Output](#output)
-   - [License](#license)
-   - [Building](#building)
-   - [Contributing](#contributing)
-
 # Frends.Community.Postgre
+
+frends Community Task for PostgreOperations
+
+[![Actions Status](https://github.com/CommunityHiQ/Frends.Community.Postgre/workflows/PackAndPushAfterMerge/badge.svg)](https://github.com/CommunityHiQ/Frends.Community.Postgre/actions) ![MyGet](https://img.shields.io/myget/frends-community/v/Frends.Community.Postgre) [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT) 
+
+- [Installing](#installing)
+- [Tasks](#tasks)
+     - [PostgreOperations](#PostgreOperations)
+- [Building](#building)
+- [Contributing](#contributing)
+- [Change Log](#change-log)
+
+# Installing
+
+You can install the Task via frends UI Task View or you can find the NuGet package from the following NuGet feed
+https://www.myget.org/F/frends-community/api/v3/index.json and in Gallery view in MyGet https://www.myget.org/feed/frends-community/package/nuget/Frends.Community.Postgre
+
+# Tasks
+
+## PostgreOperations
+
 FRENDS Task for Postgre queries.
 
-## Installing
-You can install the task via Frends UI Task view or you can find the nuget package from the following nuget feed
-https://www.myget.org/F/frends/api/v3/index.json
-
-## Tasks
-
-### QueryData
-
-#### QueryParameters
+### QueryParameters
 
 | Property          |  Type   | Description								| Example                     |
 |-------------------|---------|-----------------------------------------|-----------------------------|
-| Query				| string	| Postgre query string	| ´SELECT * FROM table WHERE \"Column\" = ''Value''´ |
+| Query				| string	| Postgre query string	| Â´SELECT * FROM table WHERE \"Column\" = ''Value''Â´ |
 | Parameters		| Array(string,string)	| List of inputs parameters	| `input` `message` |
 | ReturnType		| XMLString, JSONString	| Query return type	| JSONString |
 | ReturnType		| string | Specify the culture info to be used when parsing result to JSON and to XML. If this is left empty InvariantCulture will be used. List of cultures: https://msdn.microsoft.com/en-us/library/ee825488(v=cs.20).aspx Use the Language Culture Name.	| en-US |
 
-
-
-
-
-#### ConnectionInformation
+### ConnectionInformation
 
 | Property          |  Type   | Description								| Example                     |
 |-------------------|---------|-----------------------------------------|-----------------------------|
-| ConnectionString	  | string	| Postgre connection string | ´Server=myServerAddress;Database=myDataBase;User Id=myUsername;Password=myPassword;´ |
+| ConnectionString	  | string	| Postgre connection string | Â´Server=myServerAddress;Database=myDataBase;User Id=myUsername;Password=myPassword;Â´ |
 | TimeoutSeconds	  | int		| Timeout in seconds		| 30 |
 
-#### Output
+
+### Output
 
 | Property          |  Type   | Description								| Example                     |
 |-------------------|---------|-----------------------------------------|-----------------------------|
 | Result        | string   | Depends on the value chosen in Input	| XMLString |
 
-## License
 
-This project is licensed under the MIT License - see the LICENSE file for details
 
-## Building
-Ensure that you have https://www.myget.org/F/frends/api/v3/index.json added to your nuget feeds
 
-Clone a copy of the repo
+# Building
 
-git clone https://github.com/CommunityHiQ/Frends.Community.Postgre.git
+Clone a copy of the repository
 
-Restore dependencies
-
-nuget restore Frends.Community.Postgre
+`git clone https://github.com/CommunityHiQ/Frends.Community.Postgre.git`
 
 Rebuild the project
 
-Run Tests with nunit3. Tests can be found under
+`dotnet build`
 
-Frends.Community.Postgre.Tests\bin\Release\Frends.Community.Postgre.Tests.dll
+Run tests
 
-Create a nuget package
+`dotnet test`
 
-`nuget pack nuspec/Frends.Community.Postgre.nuspec`
+Create a NuGet package
 
-## Contributing
+`dotnet pack --configuration Release`
+
+# Contributing
 When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owners of this repository before making a change.
 
-1. Fork the repo on GitHub
+1. Fork the repository on GitHub
 2. Clone the project to your own machine
 3. Commit changes to your own branch
 4. Push your work back up to your fork
@@ -90,4 +85,5 @@ NOTE: Be sure to merge the latest from "upstream" before making a pull request!
 | 1.0.1 | Implemented ToJson and ToXml extensions, refactored main code, removed return types 'XDocument', 'XmlDocument' and 'Dynamic', updated documentation and added license information |
 | 1.1.0 | Updated documentation and fixed nuget dependencies. First pubic release. |
 | 1.2.0 | Fixed nuget dependencies again. |
+| 1.2.1 | Converted to support .Net Standard and .Net Framework 4.7.1
